feat:  DTO와 컨트롤러 수정

+++ secretboard/src/main/java/com/siat/secretboard/board/dto/BoardRequestDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BoardRequestDTO {
-    private int boardIdx;
+    private Long boardIdx;
     private String boardName;
     private String regDate;
     private String updateDate;
